requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
whitenoise>=6.6
gunicorn>=21.0
dj-database-url>=2.1
psycopg2-binary>=2.9
python-decouple>=3.8
